firebaseの初期化処理を適切な箇所に移動(firebase.client.ts)。
